% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainValid.R
\name{aggregateByDay}
\alias{aggregateByDay}
\title{aggregate rain data by day}
\usage{
aggregateByDay(x, FUN, ..., timeColumn = 1, excludeColumns = 1:2)
}
\arguments{
\item{x}{data frame containing at least one date and time column}

\item{FUN}{aggregation function}

\item{...}{arguments passed to FUN}

\item{timeColumn}{number of the date and time column}

\item{excludeColumns}{vector of column numbers to be excluded before aggregation}
}
\description{
aggregate rain data by day
}
