% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainValidSimple.R
\name{checkForDuplicates}
\alias{checkForDuplicates}
\title{Stop if there are Duplicates and show them}
\usage{
checkForDuplicates(x, keys)
}
\arguments{
\item{x}{data frame}

\item{keys}{columns that are expected to be key columns and that should not
contain duplicated value combinations}
}
\description{
Stop if there are Duplicates and show them
}
