% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainValidSimple.R
\name{compactForGauges}
\alias{compactForGauges}
\title{Compact Signal Data}
\usage{
compactForGauges(
  signals,
  gauges,
  validation,
  level = 2,
  stars = TRUE,
  dbg = TRUE
)
}
\arguments{
\item{signals}{data frame as returned by
\code{kwb.read:::.toLongFormat(signals, type = "signal")}}

\item{gauges}{character vector of rain gauge names}

\item{validation}{data frame}

\item{level}{level of compression}

\item{stars}{if \code{TRUE} (default) stars are used to indicate differences}

\item{dbg}{if \code{TRUE} (default) debug messages are shown}
}
\description{
Compact Signal Data
}
