% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainValidSimple.R
\name{correctRainDiffs}
\alias{correctRainDiffs}
\title{Correct Rain Height Differences}
\usage{
correctRainDiffs(
  rainDiffs,
  userCorrections,
  gauges = names(rainDiffs)[-(1:2)],
  dbg = TRUE
)
}
\arguments{
\item{rainDiffs}{rain height differences as provided by
\code{kwb.read:::.toDifferences}}

\item{userCorrections}{user correction data as returned by
\code{\link{readUserCorrection}}}

\item{gauges}{names of rain gauges (as they appear as column names in
\code{rainDiffs}) to be considered}

\item{dbg}{if \code{TRUE} (default) debug messages are shown}
}
\description{
Add one day to the data in the "Date" column if the time in column "To" is
before the time in column "From"
}
