% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainConvert.R
\name{corrToLongFormat}
\alias{corrToLongFormat}
\title{rain correction data from 'wide' format to 'long' format}
\usage{
corrToLongFormat(corrData)
}
\arguments{
\item{corrData}{correction data as returned by
\code{kwb.read::readBwbRainCorrection}}
}
\description{
rain correction data from 'wide' format to 'long' format
}
