% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainValidSimple.R
\name{defaultArgumentAssignment}
\alias{defaultArgumentAssignment}
\title{Output Template for Argument Assignments}
\usage{
defaultArgumentAssignment(x, type = c("R", "csv")[1])
}
\arguments{
\item{x}{vector of character}

\item{type}{one of \code{"R", "csv"}}
}
\description{
Output Template for Argument Assignments
}
