% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainValid.R
\name{getCorrectionCases}
\alias{getCorrectionCases}
\title{get the cases of required rain data correction}
\usage{
getCorrectionCases(corrData, rainData, ...)
}
\arguments{
\item{corrData}{correction data as returned by
\code{kwb.read::readBwbRainCorrection}}

\item{rainData}{rain data as returned by \code{kwb.read::readBwbRainData}}

\item{...}{additional arguments passed to the final \code{merge}, such as
\code{all.x} (keep correction data even for days for which no rain data
is available), \code{all.y} (keep rain data even for days for which no
correction data is available)}
}
\description{
get a data frame with each row representing a gauge and a day at which
rain data needs to be corrected
}
