% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainMain.R
\name{getDailyCumulativeRain}
\alias{getDailyCumulativeRain}
\title{cumulate rain data within each day}
\usage{
getDailyCumulativeRain(rd, to.long = TRUE)
}
\arguments{
\item{rd}{data frame containing rain data as returned by
\code{kwb.read::readBwbRainData}}

\item{to.long}{if \code{TRUE} (default) the result data frame is converted to
'long' format.}
}
\description{
cumulate rain data within each day
}
