% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainValidSimple.R
\name{getNeighbours}
\alias{getNeighbours}
\title{Get the Names of Neighbouring Rain Gauges}
\usage{
getNeighbours(signals, gaugeInfo, n = 2, gaugePos = NULL)
}
\arguments{
\item{signals}{data frame as returned by
\code{kwb.read:::.toLongFormat(signals, type = "signal")}}

\item{gaugeInfo}{data frame as returned by
\code{\link[kwb.read]{readAllBwbSignals}} in attribute \code{gaugeInfo}}

\item{n}{number of neighbours}

\item{gaugePos}{data frame containing the gauge positions as required by
\code{\link[kwb.read]{getGaugeDistances}}}
}
\description{
Get the Names of Neighbouring Rain Gauges
}
