% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainPlot.R
\name{plotCumulativeRain}
\alias{plotCumulativeRain}
\title{plot cumulative rain}
\usage{
plotCumulativeRain(rd.long, to.pdf = TRUE)
}
\arguments{
\item{rd.long}{data frame in 'long' format as returned by
\code{\link{getDailyCumulativeRain}}}

\item{to.pdf}{If \code{TRUE} (default) the plot goes into a temporary PDF
file, otherwise to the current plot device}
}
\description{
plot cumulative rain
}
