% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainPlot.R
\name{plotDailyRainHeightTable}
\alias{plotDailyRainHeightTable}
\title{Print Tables of Daily Rain Heights (as Plots!)}
\usage{
plotDailyRainHeightTable(
  rd,
  landscape = FALSE,
  cex = ifelse(landscape, 0.8, 0.6),
  ppp = ifelse(landscape, 1, 2),
  to.pdf = TRUE
)
}
\arguments{
\item{rd}{data frame with time stamps in first column and rain heights per
time interval for different rain gauges in all the other columns}

\item{landscape}{passed to \code{kwb.utils::preparePdf}}

\item{cex}{passed to \code{\link[kwb.misc]{hsPrintToPlot}}}

\item{ppp}{number of plots per page}

\item{to.pdf}{if \code{TRUE} the output goes into a temporary PDF file}
}
\description{
Print Tables of Daily Rain Heights (as Plots!)
}
