% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainPlot.R
\name{plotRainEventOverview}
\alias{plotRainEventOverview}
\title{Plot all rain events contained in rain data}
\usage{
plotRainEventOverview(rd, timeColumn = "tEnd_BWB", to.pdf = TRUE)
}
\arguments{
\item{rd}{data frame with time stamps in first column and rain heights per
time interval for different rain gauges in all the other columns}

\item{timeColumn}{name of the column containing the timestamps.
Default: "tEnd_BWB"}

\item{to.pdf}{if \code{TRUE} the output goes into a temporary PDF file}
}
\description{
Plot all rain events contained in rain data
}
