% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainConvert.R
\name{rainToLongFormat}
\alias{rainToLongFormat}
\title{rain data from 'wide' format to 'long' format}
\usage{
rainToLongFormat(rainData)
}
\arguments{
\item{rainData}{rain data as returned by \code{kwb.read::readBwbRainData}}
}
\description{
rain data from 'wide' format to 'long' format
}
