% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainValidSimple.R
\name{readUserCorrection}
\alias{readUserCorrection}
\title{Read User Correction Data from CSV File}
\usage{
readUserCorrection(
  file,
  sep = ";",
  country = "de",
  date.format = "\%d.\%m.\%Y",
  dbg = TRUE
)
}
\arguments{
\item{file}{full path to CSV file}

\item{sep}{column separator}

\item{country}{country code ("de" or "en")}

\item{date.format}{date format string}

\item{dbg}{if \code{TRUE} (default) debug messages are shown}
}
\description{
Read User Correction Data from CSV File
}
