% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainConvert.R
\name{reformat_rain_data}
\alias{reformat_rain_data}
\title{Reformat Rain Data}
\usage{
reformat_rain_data(rain)
}
\arguments{
\item{rain}{data frame with begin and end timestamps of five minute time
intervals in the first two (character) columns, rain heights of five minute
intervals in the remaining columns. Such a data frame is returned by
\code{\link{plotRainEventOverview}}}
}
\description{
Reformat rain data as returned by \code{\link{correctRainDiffs}} to the
format required by \code{\link{plotRainEventOverview}}.
}
