% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainValidSimple.R
\name{toGaugeSignals}
\alias{toGaugeSignals}
\title{Convert raw Signals and Validation Data to Valid Signals}
\usage{
toGaugeSignals(signals, validation, neighbours, digits = 3, ...)
}
\arguments{
\item{signals}{data frame as returned by
\code{kwb.read:::.toLongFormat(signals, type = "signal")}}

\item{validation}{data frame as returned by \code{\link{toValidationTable}}}

\item{neighbours}{vector of names of neighbour gauges}

\item{digits}{number of digits to which the signals are rounded}

\item{\dots}{arguments passed to \code{\link{compactForGauges}}}
}
\description{
Convert raw Signals and Validation Data to Valid Signals
}
