% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainValidSimple.R
\name{toValidationTable}
\alias{toValidationTable}
\title{Create Validation Table from Signal and Correction Data}
\usage{
toValidationTable(signals, corrections)
}
\arguments{
\item{signals}{data frame as returned by
\code{kwb.read:::.toLongFormat(signals, type = "signal")}}

\item{corrections}{data frame as returned by
\code{kwb.read:::.toLongFormat(corrections, type = "correction")}}
}
\description{
Create Validation Table from Signal and Correction Data
}
