% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainValidSimple.R
\name{writeDiffFiles}
\alias{writeDiffFiles}
\title{Write Files containing the "Diff" between raw and valid rain signal}
\usage{
writeDiffFiles(
  gaugeSignals,
  dir.out,
  gauges = names(gaugeSignals),
  years = NULL,
  suffix = NULL,
  sep = ";",
  dec = ",",
  dbg = TRUE
)
}
\arguments{
\item{gaugeSignals}{list of data frames with each element representing one
rain gauge}

\item{dir.out}{full path to output directory}

\item{gauges}{vector of character containing the names of the gauges for
which files are to be written}

\item{years}{vector of integer representing the years for which diff files
are to be generated}

\item{suffix}{to be used in the names of the created files}

\item{sep}{column separator in created CSV files}

\item{dec}{decimal character in created CSV files}

\item{dbg}{if \code{TRUE} (default) debug messages are shown}
}
\description{
Write Files containing the "Diff" between raw and valid rain signal
}
